Remove focus neighbors for subscene back buttons
For any control in menu sub-scenes that are in group "menu_button",
the focus_neighbour_left and focus_neighbour_right attributes are set
to the controls themselves so that focus won't go to the invisible
parent/sub-menus when left or right button is set.

This shouldn't have been necessary, but somehow even though those
attributes are unset, the left/right buttons work without this.